product title clickable to enter detail page
